add validation to check v1beta2 status
